Updated twitch api url.
Added extra call to twitch api /channels resource. Previously, the code only called /streams resource. If the stremaer was not online a streamer logo was unavailable. Calling /channels resource we have access to streamer logo. This provides a better UI layout with streamer logo instead of a default image from
http://placehold.it.